Add step counter and date
I had to rearrange the display and this will need to be fixed. The goal, however, was to get more data on the screen. I added the date and a step counter. I still need to animate stuff, add a battery (possibly), add the weather, fix font stuff, organize the screen, make sure that the app runs on all watchfaces, deal with color choice, and make settings.